+++ bolt/util.py
@@ -1,7 +1,9 @@
 import pathlib
 import subprocess
 import sys
+import os
 import textwrap
+import pysam
 
 
 from .common import constants
@@ -99,3 +101,88 @@ def get_qualified_vcf_annotation(anno_enum):
 #    existing_filters = [e for e in record.FILTERS if e != 'PASS']
 #    assert filter_enum.value not in existing_filters
 #    return ';'.join([*existing_filters, filter_enum.value])
+
+def split_vcf(input_vcf, output_dir, max_variants=100000000):
+    """
+    Splits a VCF file into multiple chunks, each containing up to max_variants variants.
+    Each chunk includes the VCF header.
+    """
+    # Ensure output_dir exists
+    # Count total number of variants using util.count_vcf_records
+    total_variants = count_vcf_records(input_vcf)
+    print(f"Total number of variants in the input VCF: {total_variants}")
+
+
+    chunk_files = []
+    chunk_number = 1
+    variant_count = 0
+    base_filename = os.path.splitext(os.path.basename(input_vcf))[0]
+    chunk_filename = os.path.join(output_dir, f"{base_filename}_chunk{chunk_number}.vcf")
+    chunk_files.append(chunk_filename)
+
+    # Open the input VCF using pysam
+    vcf_in = pysam.VariantFile(input_vcf, 'r')
+    # Create a new VCF file for the first chunk
+    vcf_out = pysam.VariantFile(chunk_filename, 'w', header=vcf_in.header)
+
+    for record in vcf_in:
+        if variant_count >= max_variants:
+            # Close the current chunk file and start a new one
+            vcf_out.close()
+            chunk_number += 1
+            chunk_filename = os.path.join(output_dir, f"{base_filename}_chunk{chunk_number}.vcf")
+            chunk_files.append(chunk_filename)
+            vcf_out = pysam.VariantFile(chunk_filename, 'w', header=vcf_in.header)
+            variant_count = 0
+        # Write the record to the current chunk
+        vcf_out.write(record)
+        variant_count += 1
+
+    # Close the last chunk file
+    vcf_out.close()
+    vcf_in.close()
+
+    print(f"VCF file split into {len(chunk_files)} chunks.")
+
+    return chunk_files
+
+def merge_tsv_files(tsv_files, merged_tsv_fp):
+    """
+    Merges all TSV files into a single TSV.
+    """
+    with open(merged_tsv_fp, 'w') as merged_tsv:
+        for i, tsv_file in enumerate(tsv_files):
+            with open(tsv_file, 'r') as infile:
+                for line_number, line in enumerate(infile):
+                    # Skip header except for the first file
+                    if i > 0 and line_number == 0:
+                        continue
+                    merged_tsv.write(line)
+    print(f"Merged TSV written to: {merged_tsv_fp}")
+
+
+def merge_vcf_files(vcf_files, merged_vcf_fp):
+    merged_unsorted_vcf = merged_vcf_fp + '.unsorted.vcf.gz'
+    merged_vcf = merged_vcf_fp + '.vcf.gz'
+    cmd = ['bcftools', 'merge', '-m', 'all', '-Oz', '-o', merged_unsorted_vcf] + vcf_files
+    # Run the bcftools merge command
+    try:
+        print("Running bcftools merge...")
+        subprocess.run(cmd, check=True)
+        print(f"Merged VCF written to: {merged_unsorted_vcf}")
+    except subprocess.CalledProcessError as e:
+        print(f"Error merging VCF files with bcftools:\n{e}")
+        raise
+
+    # Sort the merged VCF file
+    cmd_sort = ['bcftools', 'sort', '-Oz', '-o', merged_vcf, merged_unsorted_vcf]
+    try:
+        print(f"Sorting merged VCF file...")
+        subprocess.run(cmd_sort, check=True)
+        print(f"Sorted merged VCF written to: {merged_vcf}")
+    except subprocess.CalledProcessError as e:
+        print(f"Error sorting merged VCF file:\n{e}")
+        raise
+
+    return merged_vcf
+
